to int instead of float

+++ API.py
@@ -22,7 +22,7 @@ class transactions:
         #Query DB and get all transactions 
         transactions = db.select('relevant_transactions', where='Filing_date > $key', vars=vars)   
         transactions = [ dict(entry) for entry in transactions ]
-        transactions = [float('Total_value') for entry in transactions['Total_value']]
+        transactions = [int('Total_value') for entry in transactions['Total_value']]
         
         for item in transactions:
             item['Total_value'] = numbers.format_currency(item['Total_value'],item['Currency'], locale='es_US')
